backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-jwt-extended==4.5.2
torch==2.0.1
torchvision==0.15.2
pymongo==4.15.5
dnspython==2.7.0
python-dotenv==1.2.1
pillow==10.0.0
numpy==1.24.3
matplotlib==3.8.0
pandas==2.0.3
scikit-learn==1.3.0
requests==2.32.3